Add back resolvers folder
