+++ cli/npm/managed/resolvers/common.rs
