+++ cli/npm/managed/resolvers/global.rs
@@ -22,11 +22,11 @@ use deno_runtime::deno_node::NodeResolutionMode;
 
 use crate::npm::NpmCache;
 
+use super::super::resolution::NpmResolution;
 use super::common::cache_packages;
 use super::common::types_package_name;
 use super::common::NpmPackageFsResolver;
 use super::common::RegistryReadPermissionChecker;
-use super::resolution::NpmResolution;
 
 /// Resolves packages from the global npm cache.
 #[derive(Debug)]
